move reused function to helpers
