+++ sites/geohub/src/components/controls/RasterIntervalsLegend.svelte
@@ -22,6 +22,7 @@
   import IntervalsLegendColorMapRow from '$components/controls/IntervalsLegendColorMapRow.svelte'
   import type { IntervalLegendColorMapRow, Layer, RasterLayerStats, RasterTileMetadata } from '$lib/types'
   import { layerList, map } from '$stores'
+  import {updateIntervalValues} from "$lib/helper/updateIntervalValues";
 
   export let colorPickerVisibleIndex: number
   export let layerConfig: Layer
@@ -169,66 +170,11 @@
     colorPickerVisibleIndex = event.detail.index
   }
   const handleChangeIntervalValues = (event: CustomEvent) => {
-    const rowIndex = event.detail.index
-    const inputType = event.detail.id
-    let inputValue = event.detail.value as number
-    let currentRow = colorMapRows.at(rowIndex)
-    if (rowIndex == 0) {
-      const nextRow = colorMapRows.at(rowIndex + 1)
-      if (inputType == 'start') {
-        inputValue =
-          !isNaN(inputValue) && inputValue < currentRow.end && inputValue >= currentRow.start
-            ? inputValue
-            : (currentRow.start as number)
-        colorMapRows[rowIndex].start = inputValue
-      } else {
-        inputValue =
-          !isNaN(inputValue) && inputValue > currentRow.start && inputValue < nextRow.end
-            ? inputValue
-            : (currentRow.end as number)
-        colorMapRows[rowIndex].end = inputValue
-        colorMapRows[rowIndex + 1].start = inputValue
-      }
-    } else if (rowIndex == colorMapRows.length - 1) {
-      const prevRow = colorMapRows.at(rowIndex - 1)
-      if (inputType == 'start') {
-        inputValue =
-          !isNaN(inputValue) && inputValue < currentRow.end && inputValue > prevRow.start
-            ? inputValue
-            : (currentRow.start as number)
-        colorMapRows[rowIndex].start = inputValue
-        colorMapRows[rowIndex - 1].end = inputValue
-      } else {
-        inputValue =
-          !isNaN(inputValue) && inputValue <= currentRow.end && inputValue > prevRow.start
-            ? inputValue
-            : (currentRow.end as number)
-        colorMapRows[rowIndex].end = inputValue
-      }
-    } else {
-      const nextRow = colorMapRows.at(rowIndex + 1)
-      const prevRow = colorMapRows.at(rowIndex - 1)
-
-      if (inputType == 'start') {
-        inputValue =
-          !isNaN(inputValue) && inputValue > prevRow.start && inputValue < currentRow.end
-            ? inputValue
-            : (currentRow.start as number)
-        colorMapRows[rowIndex].start = inputValue
-        colorMapRows[rowIndex - 1].end = inputValue
-      } else {
-        inputValue =
-          !isNaN(inputValue) && inputValue > currentRow.start && inputValue < nextRow.end
-            ? inputValue
-            : (currentRow.end as number)
-        colorMapRows[rowIndex].end = inputValue
-        colorMapRows[rowIndex + 1].start = inputValue
-      }
-    }
-
+    colorMapRows = updateIntervalValues(event, colorMapRows)
     rowWidth = getMaxValueOfCharsInIntervals(colorMapRows)
     handleParamsUpdate()
   }
+
 </script>
 
 <div
